voucher

+++ database/migrations/2023_04_07_001914_create_origins_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('tblOrigin', function (Blueprint $table) {
-            $table->bigIncrements('OrgID');
+            $table->integer('OrgID');
             $table->string('OrgName')->unique();
             $table->timestamps();
         });
